I noticed that my programs were not linking with the shared libs.  Apparently if you say something like (from the dmapi tests stuff):
	gcc -g -O2 -o dm_handle dm_handle.o libdmtest.a -ldm -lhandle

it will link with whatever it finds in /usr/lib.  If there are no shared libs
there then it'll link with the static libs.  When I add a symlink to make
libdm.so and libhandle.so appear in /usr/lib then I am able to link to the
shared libs.
No Message Supplied