dmapi: don't hard code the shell binary
(This is just replicating the change that Dave Chinner made in
xfsprogs, xfsdump, and xfstests.)

Recent changes to debian unstable user space have caused the
xfsdump build to break as certain shell functionality is being
assumed by libtool and friends.  The configure scripts test and
select the correct shell, but the input files ignore this and hard
code the shell to use and hence now break.

Fix this by using the shell that the configure scripts decide is the
right one to use.

Signed-off-by: Alex Elder <aelder@sgi.com>